remove float upcast in utils.resample, use boolean dtype conditioning mask in unet trainer, add logging of per row latents std in ddec trainer i1, dae / ddec i1 tweaks (add conv2d reg for combined latents)

+++ src/utils/resample.py
@@ -54,11 +54,11 @@ class FilteredResample1D(torch.nn.Module):
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
 
-        kernel: torch.Tensor = self.kernel.float()
+        kernel: torch.Tensor = self.kernel
         original_dtype = x.dtype
 
         kw = kernel[None, None, None, :].expand(x.shape[1], 1, 1, self.k_size)
-        x = torch.nn.functional.conv2d(self.pad_w(x).float(), kw, groups=x.shape[1], stride=(1,self.stride))
+        x = torch.nn.functional.conv2d(self.pad_w(x), kw, groups=x.shape[1], stride=(1,self.stride))
 
         return x.to(dtype=original_dtype)
 
@@ -111,7 +111,7 @@ class FilteredResample2D(torch.nn.Module):
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
 
-        kernel: torch.Tensor = self.kernel.float()
+        kernel: torch.Tensor = self.kernel
 
         kw = kernel[None, None, None, :].expand(x.shape[1], 1, 1, self.k_size)
         x = torch.nn.functional.conv2d(self.pad_w(x), kw, groups=x.shape[1], stride=(1,self.stride))
@@ -245,11 +245,11 @@ class FilteredResample1D3(torch.nn.Module):
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
 
-        kernel: torch.Tensor = self.kernel.float()
+        kernel: torch.Tensor = self.kernel
         original_dtype = x.dtype
 
         kw = kernel[None, None, None, None, :].expand(x.shape[1], 1, 1, 1, self.k_size)
-        x = torch.nn.functional.conv3d(self.pad_w(x).float(), kw, groups=x.shape[1], stride=(1,1,self.stride))
+        x = torch.nn.functional.conv3d(self.pad_w(x), kw, groups=x.shape[1], stride=(1,1,self.stride))
 
         return x.to(dtype=original_dtype)
